Revert "[Tizen]Change the precision of size and uSize from mediumP to highP"
This reverts commit 4036d3983f42718568661a2059c272b342397c1f.

+++ dali-toolkit/internal/controls/video-view/video-view-impl.cpp
@@ -88,7 +88,7 @@ const char* const CUSTOM_SAMPLER_TYPE_NAME( "samplerExternalOES" );
 const char* VERTEX_SHADER = DALI_COMPOSE_SHADER(
   attribute mediump vec2 aPosition;\n
   uniform highp   mat4 uMvpMatrix;\n
-  uniform highp   vec3 uSize;\n
+  uniform mediump vec3 uSize;\n
   \n
   void main()\n
   {\n
@@ -110,7 +110,7 @@ const char* VERTEX_SHADER_TEXTURE = DALI_COMPOSE_SHADER(
   attribute mediump vec2 aPosition;\n
   varying mediump vec2 vTexCoord;\n
   uniform highp   mat4 uMvpMatrix;\n
-  uniform highp   vec3 uSize;\n
+  uniform mediump vec3 uSize;\n
   varying mediump vec2 sTexCoordRect;\n
   void main()\n
   {\n
